chore: minor refactor and tweaks in prep for PyPI pub
