+++ python/README.md
@@ -1,0 +1,3 @@
+# Argo Client
+
+A python library for connecting to RPC servers, specifically desigend with the Haskell [Argo]([https://](https://github.com/galoisinc/argo)) sister library in mind.
