+++ python/setup.cfg
@@ -1,0 +1,20 @@
+[metadata]
+name = argo_client
+version = 0.0.1
+url = https://github.com/GaloisInc/argo/python
+author = Galois Inc.
+author_email = andrew@galois.com
+license = BSD 3-Clause License
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: BSD License
+    Operating System :: OS Independent
+description = An RPC client library
+long_description = file: README.md
+long_description_content_type = text/markdown
+
+[options]
+python_requires = >=3.7
+install_requires =
+    requests
+    mypy
