+++ python/setup.py
@@ -1,9 +1,3 @@
-from setuptools import setup, find_packages
+import setuptools
 
-setup(name='argo',
-      version='0.0.1',
-      packages=['argo'],
-      package_data={
-          '': ['py.typed']
-      }
-)
+setuptools.setup()
